Possibly add line break when processing wordwrap breakpoint
When word wrapping: Before a word plus its trailing separator
(breakpoint) is added to a line, a check ensures that adding both will
not exceed the limit. Without this check, it is possible that both are
added when adding the word alone would exceed the limit, but adding both
does. This would result in the line exceeding its limit.

This also fixes scenarios where a double separator could also grow a
line beyond the limit.